feat: docker images
This adds a dockerfile + supporting CI and files to build a docker image
for dd-metric. This image is based on scratch for minimal size and uses
the appropriate musl target to build dd-metric.

The image is built for linux/amd64 and linux/arm64 platforms

This also adds more CI steps. PRs now run builds of the binary and
docker image, but do not upload them